+++ inst/translation/trees_fr.R
@@ -1,4 +1,4 @@
-.trees_en <- function(trees, labels_only = FALSE, as_labelled = FALSE) {
+.trees_fr <- function(trees, labels_only = FALSE, as_labelled = FALSE) {
   # The trees dataset has a couple of glitches:
   # 1) Girth is indeed the Diameter (and variables not in snake_case), so rename
   names(trees) <- c("diameter", "height", "volume")